Move scan to utils (#6)
* Move scan to utils

* Forgot one refrence
