+++ googledevices/utils/scan.py
